+++ chapter-1/seconds-to-years.rb
@@ -1,0 +1,19 @@
+puts ""
+puts "# = = = = = = = = = = "
+puts "# Chapter 1"
+puts "# File: seconds-to-years.rb"
+puts "# = = = = = = = = = = "
+puts ""
+
+puts "Constants:"
+puts "Months within year: 12"
+puts "Days within year: 365"
+puts "Hours within day: 24"
+puts "Minutes within hour: 60"
+puts "Seconds within minute: 60"
+puts "Seconds during life: 1406000000"
+puts ""
+
+puts "Life in years:"
+puts 1406000000/60/60/24/365
+puts ""