+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - minutes-within-decade.rb [todo from chapter-1]
 - seconds-old.rb [todo from chapter-1]
 - chocolates-within-life.rb [todo from chapter-1]
+- seconds-to-years.rb [todo from chapter-1]